+++ common_test.go
@@ -2,7 +2,6 @@ package lr0
 
 import (
 	"strconv"
-	"unicode"
 
 	"github.com/pkg/errors"
 )
@@ -38,36 +37,6 @@ func matchIdentifier(state *State) (next *State, value any) {
 	return
 }
 
-func matchDigitsStr(state *State) (next *State, value any) {
-	st, b := state.TakeBytesFunc(isDigit)
-	if b == nil {
-		return
-	}
-	return st, string(state.BytesTo(st))
-}
-func matchDigits(state *State) (next *State, value any) {
-	next, value = matchDigitsStr(state)
-	if next == nil {
-		return
-	}
-	if _, ok := value.(error); ok {
-		return
-	}
-	value, err := strconv.Atoi(string(state.BytesTo(next)))
-	if err != nil {
-		value = err
-	}
-	return
-}
-
-func matchWS(st *State) (next *State, v any) {
-	to, _ := st.TakeRunesFunc(unicode.IsSpace)
-	if to.Offset() == st.Offset() {
-		return nil, nil
-	}
-	return to, nil
-}
-
 func isDigit(b byte) bool        { return b >= '0' && b <= '9' }
 func byteIsNotSpace(b byte) bool { return b != ' ' }
 func runeIsNotSpace(r rune) bool { return r != ' ' }
@@ -81,6 +50,8 @@ func isAlpha(b byte) bool {
 		return false
 	}
 }
+
+func bytesToInt(b []byte) (int, error) { return strconv.Atoi(string(b)) }
 
 func calc3AnyNil(any, any, any) (any, error)     { return nil, nil }
 func calc3StrTrace(a, op, b string) (any, error) { return "(" + a + " " + op + " " + b + ")", nil }